[red-knot] Move `Ty` enum to property tests
